apis/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==4.2.16
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-dbbackup==5.0.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.61.1
frozenlist==1.8.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
multidict==6.7.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pgvector==0.4.2
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rest-framework-simplejwt==0.0.2
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
twilio==9.9.1
tzdata==2025.2
urllib3==2.6.3
wheel==0.45.1
yarl==1.22.0
dlib==20.0.0
face_recognition==1.3.0
face_recognition_models==0.3.0
